+++ src/Shapes/Triangle.cs
@@ -13,7 +13,7 @@ namespace ShapeEditor.Shapes
                         Point point3,
                         Color borderColor,
                         Color fillColor,
-                        double borderWidth) : this(point1, point2, point3)
+                        float borderWidth) : this(point1, point2, point3)
         {
             BorderColor = borderColor;
             FillColor = fillColor;
@@ -30,12 +30,12 @@ namespace ShapeEditor.Shapes
 
         public void Draw(IRenderer render)
         {
-            render.FillPolygon(Points, BorderColor, FillColor,(float) BorderWidth);
+            render.FillPolygon(Points, BorderColor, FillColor, BorderWidth);
         }
 
         public Color FillColor { get; set; }
         public Color BorderColor { get; set; }
-        public double BorderWidth { get; set; }
+        public float BorderWidth { get; set; }
         public IEnumerable<Point> Points { get; }
         public bool IsInside(Point point)
         {
